feat: get accessToken for Kitsu (#33)
* test: remove unused test

Remove test file that is was made during project creation.

* feat: begin adding kitsu route

Begin adding route to communicate with Kitsu.

* style: add ignore for import ordering

Import ordering rule doesnt look to be working, will look to fix it later on.

* feat: add basics to send Kitsu auth data to Orochi

Add the required model and serialization so that we can send data into Orochi.

* refactor: move packages around so its easier to see the routes

Make the routing package just used for setting up the routing, and move the routes out into their
own routes package.

* feat: add successful request to Kitsu auth

Can now make a successful request to the kitsu auth endpoint. This will currently just get the
response but will not parse it out into a valid model for returning or displaying back to a
consumer.

* refactor: cleanup the packages slightly

Move some files around so they fit into packages that are more correct for their purpose.

* feat: return model from the auth endpoint

When getting a success from the Kitsu auth endpoint, return a model containing the data.

* refactor: return custom domain on successful auth

Return custom domain instead of the dto when a successful auth request occurs.

* chore: add explicit reflect dependency

Reflect is needed to ensure it runs on the same kotlin version as other kotlin jars

* test: fix test missing HttpClient

* style: update detekt rules to use new syntax

Detekt rules updated from using a string for lists to using an array.