refactor: add service to route (#49)
* refactor: move logic from the route into a service class

Move out from the kitsuAuth route, and instead put the logic into a KitsuAuthService. As part of
this implemented a result library to make calling it a bit nicer.

* refactor: move creation of the response models to the service

Move away from the route so that the service does a bit more work, keeping the route more clean.

* test: begin writing tests for the KitsuAuthService

* test: add tests for the KitsuAuthService

Add some unit tests around the KitsuAuthService. Unfortuantely due to mockk not being able to mock
methods that are reified, the calls to the API need to use the MockEngine instead of simply mocking
some return values.

* refactor: move routing from kitsu/auth to auth/kitsu

Move to auth-kitsu so that all the auth packages can be put together, instead of the different
services put together.

* test: add integration test for the kitsu auth route

Add integration test to cover off the happy path for now.

* test: make it easier to share models in tests

Add the models to a shared file so they aren't repeated across all of the tests.

* feat: add the StatusPage plugin

When we receive invalid JSON either from the input request, or from the API response we need to
handle this gracefully. To do this use the StatusPage plugin which lets us catch all unhandled
exceptions and return a default error domain.

close #36

* test: add integration tests for KitsuAuth failure states

* test: add helper method to start the test application

Add method that starts the test application with the same modules as the production application,
reducing the risk of misconfiguring a test.

* style: fix various detekt warnings

Add some missing kdoc in various locations, and other minor cleanups.