feat: pull down the kitsu series into a model object
Parse out the Kitsu series response into a model object instead of returning it as a string. This
will allow us to make any modifications we need to, and return as a generic data type from the api.
To allow for missing keys a new Json instance was created call Kson, which will be used to perform
the manual deserialization.